+++ autozynq/prisma/migrations/20260102154452_execution_engine_v1/migration.sql
@@ -1,0 +1,5 @@
+-- AlterTable
+ALTER TABLE "Execution" ADD COLUMN     "error" JSONB,
+ADD COLUMN     "result" JSONB,
+ADD COLUMN     "steps" JSONB,
+ADD COLUMN     "userId" TEXT;